wallet frontend: adapt unit tests for merged components - add tests to ContributionMessagesListItem.spec.js - remove slots tests
